@@ -11,31 +11,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "14.1.3",
-    "@angular/cdk": "14.1.3",
-    "@angular/common": "14.1.3",
-    "@angular/compiler": "14.1.3",
-    "@angular/core": "14.1.3",
-    "@angular/flex-layout": "14.0.0-beta.40",
-    "@angular/forms": "14.1.3",
-    "@angular/material": "14.1.3",
-    "@angular/platform-browser": "14.1.3",
-    "@angular/platform-browser-dynamic": "14.1.3",
-    "@angular/router": "14.1.3",
-    "ngx-material-file-input": "4.0.0",
+    "@angular/animations": "20.0.6",
+    "@angular/cdk": "^20.0.5",
+    "@angular/common": "20.0.6",
+    "@angular/compiler": "20.0.6",
+    "@angular/core": "20.0.6",
+    "@angular/forms": "20.0.6",
+    "@angular/material": "^20.0.5",
+    "@angular/platform-browser": "20.0.6",
+    "@angular/platform-browser-dynamic": "20.0.6",
+    "@angular/router": "20.0.6",
+    "ng-flex-layout": "^17.3.7-beta.1",
+    "ngx-custom-material-file-input": "^19.0.1",
     "rxjs": "7.5.6",
     "tslib": "2.4.0",
-    "zone.js": "0.11.8"
+    "zone.js": "^0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "14.1.3",
     "@angular-eslint/builder": "14.0.2",
     "@angular-eslint/eslint-plugin": "14.0.2",
     "@angular-eslint/eslint-plugin-template": "14.0.2",
-    "@angular-eslint/schematics": "14.0.2",
+    "@angular-eslint/schematics": "^18.4.3",
     "@angular-eslint/template-parser": "14.0.2",
-    "@angular/cli": "14.1.3",
-    "@angular/compiler-cli": "14.1.3",
+    "@angular/build": "^20.0.5",
+    "@angular/cli": "20.0.5",
+    "@angular/compiler-cli": "20.0.6",
     "@types/jasmine": "4.0.3",
     "@typescript-eslint/eslint-plugin": "5.33.1",
     "@typescript-eslint/parser": "5.33.1",
@@ -46,6 +46,6 @@
     "karma-coverage": "2.2.0",
     "karma-jasmine": "5.1.0",
     "karma-jasmine-html-reporter": "2.0.0",
-    "typescript": "4.7.4"
+    "typescript": "5.8.3"
   }
 }
