@@ -4,9 +4,10 @@ import { SessionService } from 'src/app/services/session.service';
 import { RentalsService } from '../../services/rentals.service';
 
 @Component({
-  selector: 'app-list',
-  templateUrl: './list.component.html',
-  styleUrls: ['./list.component.scss']
+    selector: 'app-list',
+    templateUrl: './list.component.html',
+    styleUrls: ['./list.component.scss'],
+    standalone: false
 })
 export class ListComponent {
 
