@@ -1,9 +1,9 @@
 import {Injectable} from "@angular/core";
-import {CanActivate, Router} from "@angular/router"; 
+import { Router } from "@angular/router"; 
 import { SessionService } from "../services/session.service";
 
 @Injectable({providedIn: 'root'})
-export class AuthGuard implements CanActivate {
+export class AuthGuard  {
 
   constructor( 
     private router: Router,
